refactor: Value -> AppSpan
