+++ src/components/AppSpan/index.js
@@ -1,0 +1,3 @@
+import AppSpan from './AppSpan';
+
+export { AppSpan as default, AppSpan };
