+++ src/components/Tag/index.js
@@ -1,0 +1,4 @@
+import Tag from './Tag';
+import TagCloud from './TagCloud';
+
+export { Tag, TagCloud };
