--- src/components/Values/index.js
@@ -1,5 +1,0 @@
-import Value from './Value';
-import Tag from './Tag';
-import TagCloud from './TagCloud';
-
-export { Value, Tag, TagCloud };
