+++ src/components/index.js
@@ -1,34 +1,33 @@
-/**
- * Components used more then once
- */
+// Tools and templates
+import withSuspense from './withSuspense';
 import ErrorBoundary from './ErrorBoundary';
 import NotImplemented from './NotImplemented';
-import withSuspense from './withSuspense';
+import ListObjectProps from './ListObjectProps';
+// App styled Components
 import AppLink from './AppLink';
 import AppIcon from './AppIcon';
 import AppButton from './AppButton';
 import AppIconButton from './AppIconButton';
 import AppAlert from './AppAlert';
 import AppSection from './AppSection';
-
-import ListObjectProps from './ListObjectProps';
+import AppSpan from './AppSpan';
+// Other components
+import { Tag, TagCloud } from './Tag';
 import FinalMessage from './FinalMessage';
-
-import { Value, Tag, TagCloud } from './Values';
 
 export {
   withSuspense,
-  NotImplemented,
   ErrorBoundary,
+  NotImplemented,
+  ListObjectProps,
   AppLink,
   AppIcon,
   AppButton,
   AppIconButton,
   AppAlert,
   AppSection,
-  Value,
+  AppSpan,
   Tag,
   TagCloud,
-  ListObjectProps,
   FinalMessage,
 };
